+++ trixnity-messenger-compose-view/src/androidMain/kotlin/de/connect2x/messenger/compose/view/files/getClipboardFile.android.kt
@@ -3,6 +3,6 @@ package de.connect2x.messenger.compose.view.files
 import de.connect2x.trixnity.messenger.util.FileDescriptor
 import okio.FileSystem
 
-actual fun getClipboardFile(fileSystem: FileSystem): Result<FileDescriptor?> {
+actual fun getClipboardFile(fileSystem: FileSystem, maxAttachmentSize: Long): Result<FileDescriptor?> {
     return Result.success(null)
 }