+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/files/PdfOverlay.kt
@@ -2,6 +2,7 @@ package de.connect2x.messenger.compose.view.files
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.focusable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
@@ -11,9 +12,11 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Button
 import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -28,10 +31,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
+import de.connect2x.messenger.compose.view.buttonPointerModifier
 import de.connect2x.messenger.compose.view.common.DownloadProgress
 import de.connect2x.messenger.compose.view.common.blockPointerInput
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
+import de.connect2x.messenger.compose.view.theme.messengerIcons
 import de.connect2x.trixnity.messenger.viewmodel.files.PdfDocumentViewModel
 import kotlin.math.ceil
 import kotlin.math.max
@@ -52,8 +57,9 @@ class PdfOverlayViewImpl : PdfOverlayView {
     @Composable
     @OptIn(ExperimentalLayoutApi::class)
     override fun create(documentViewModel: PdfDocumentViewModel) {
-        val media = documentViewModel.mediaDataFlow.collectAsState()
-        val progress = documentViewModel.progress.collectAsState()
+        val media = documentViewModel.documentFlow.collectAsState()
+        val progress = documentViewModel.progress.collectAsState().value
+        val error = documentViewModel.error.collectAsState().value
         var zoom by remember { mutableStateOf(1.0f) }
         val i18n = DI.current.get<I18nView>()
         BoxWithConstraints {
@@ -68,16 +74,22 @@ class PdfOverlayViewImpl : PdfOverlayView {
                     FlowRow {
                         Row(verticalAlignment = Alignment.CenterVertically) {
                             Button(
-                                modifier = Modifier.padding(horizontal = 8.dp),
+                                modifier = Modifier.padding(horizontal = 8.dp).buttonPointerModifier(),
                                 onClick = { documentViewModel.closeMedia() },
                             ) {
                                 Text(i18n.commonClose())
                             }
                             Text(documentViewModel.fileName)
                         }
+                        Button(
+                            modifier = Modifier.padding(horizontal = 8.dp).buttonPointerModifier(),
+                            onClick = { documentViewModel.downloadMedia() }
+                        ) {
+                            Text(i18n.downloadMessage())
+                        }
                         Row(verticalAlignment = Alignment.CenterVertically) {
                             Button(
-                                modifier = Modifier.padding(horizontal = 8.dp),
+                                modifier = Modifier.padding(horizontal = 8.dp).buttonPointerModifier(),
                                 onClick = {
                                     val newZoom = min(4.0f, zoom * 1.33f)
                                     zoom = if (newZoom > 1f && zoom < 1f) 1f else newZoom
@@ -86,7 +98,7 @@ class PdfOverlayViewImpl : PdfOverlayView {
                             }
                             Text("${ceil(zoom * 100)}%")
                             Button(
-                                modifier = Modifier.padding(horizontal = 8.dp),
+                                modifier = Modifier.padding(horizontal = 8.dp).buttonPointerModifier(),
                                 onClick = {
                                     val newZoom = max(0.1f, zoom * 0.66f)
                                     zoom = if (newZoom < 1f && zoom > 1f) 1f else newZoom
@@ -104,9 +116,25 @@ class PdfOverlayViewImpl : PdfOverlayView {
                         .weight(1f)
                         .focusable()
                 ) {
-                    if (media.value != null) PDFReader(documentViewModel, zoom)
-                    else progress.value?.let {
-                        DownloadProgress(it, documentViewModel::cancelMediaDownload)
+                    if (media.value != null) {
+                        PDFReader(documentViewModel, zoom)
+                    } else if (progress != null) {
+                        DownloadProgress(progress, documentViewModel::cancelMediaDownload)
+                    } else {
+                        Column(
+                            horizontalAlignment = Alignment.CenterHorizontally,
+                            verticalArrangement = Arrangement.Center,
+                            modifier = Modifier.fillMaxSize().padding(32.dp),
+                        ) {
+                            Icon(
+                                MaterialTheme.messengerIcons.typeFile,
+                                i18n.commonFile(),
+                                Modifier.size(96.dp).align(Alignment.CenterHorizontally)
+                            )
+                            if (error != null) {
+                                Text(error)
+                            } else Text(i18n.fileCouldNotBeLoaded())
+                        }
                     }
                 }
             }