+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/RoomViewModelTest.kt
@@ -415,7 +415,7 @@ class RoomViewModelTest : ShouldSpec() {
             ),
             roomId = roomId,
             onRoomBack = mock(),
-            onOpenModal = mock(),
+            onOpenMedia = mock(),
             isBackButtonVisible = MutableStateFlow(false),
             onOpenAvatarCutter = { _, _, _ -> },
             onOpenMention = mock(),