+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/ImageMessageViewModelTest.kt
@@ -148,7 +148,7 @@ class ImageMessageViewModelTest : ShouldSpec() {
             showSender = MutableStateFlow(true),
             sender = MutableStateFlow(UserInfoElement("User1", UserId("user1:localhost"))),
             invitation = flowOf(null),
-            onOpenModal = mock(),
+            onOpenMedia = mock(),
             mediaUploadProgress = MutableStateFlow(null),
         )
     }