+++ CHANGELOG.md
@@ -13,11 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make Smoketests work on Web
 - Make Smoketests work on Android
 - Added hook on Android to change the behaviour of the messenger on startup of the Activity
+- Add download button for file overlays
 
 ### Changed
 
 - Do not run kmmPublish on push to main
 - Automatically proceed bootstrapping without next button
+- Open video player overlay when clicking on the preview
+- Download text and markdown files when clicking on them in the timeline
 
 ### Deprecated